load_mon: split out header
